Revert some changes

+++ src/scheduler/affinity.rs
@@ -2,11 +2,11 @@ use super::worker::ThreadId;
 use crate::util::options::AffinityKind;
 
 /// Represents the ID of a logical CPU on a system.
-pub type CoreId = usize;
+pub type CoreId = u16;
 
 /// Return the total number of cores allocated to the program.
-pub fn get_total_num_cpus() -> usize {
-    core_affinity::get_core_ids().unwrap().len()
+pub fn get_total_num_cpus() -> u16 {
+    core_affinity::get_core_ids().unwrap().len() as u16
 }
 
 impl AffinityKind {
@@ -27,7 +27,7 @@ impl AffinityKind {
 
 /// Bind the current thread to the specified core.
 fn bind_current_thread_to_core(cpu: CoreId) {
-    if !core_affinity::set_for_current(core_affinity::CoreId { id: cpu }) {
+    if !core_affinity::set_for_current(core_affinity::CoreId { id: cpu as usize }) {
         panic!("Failed to bind current thread to {:?}", cpu);
     };
 }
